Add WithCancellation for async enumerables
