--- src/System.Private.CoreLib/shared/System/Runtime/CompilerServices/ConfiguredAsyncEnumerable.cs
@@ -1,48 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Threading.Tasks.Sources;
-
-namespace System.Runtime.CompilerServices
-{
-    [StructLayout(LayoutKind.Auto)]
-    public readonly struct ConfiguredAsyncEnumerable<T>
-    {
-        private readonly IAsyncEnumerable<T> _enumerable;
-        private readonly bool _continueOnCapturedContext;
-
-        internal ConfiguredAsyncEnumerable(IAsyncEnumerable<T> enumerable, bool continueOnCapturedContext)
-        {
-            _enumerable = enumerable;
-            _continueOnCapturedContext = continueOnCapturedContext;
-        }
-
-        public Enumerator GetAsyncEnumerator(CancellationToken cancellationToken = default) =>
-            new Enumerator(_enumerable.GetAsyncEnumerator(cancellationToken), _continueOnCapturedContext);
-
-        public readonly struct Enumerator
-        {
-            private readonly IAsyncEnumerator<T> _enumerator;
-            private readonly bool _continueOnCapturedContext;
-
-            internal Enumerator(IAsyncEnumerator<T> enumerator, bool continueOnCapturedContext)
-            {
-                _enumerator = enumerator;
-                _continueOnCapturedContext = continueOnCapturedContext;
-            }
-
-            public ConfiguredValueTaskAwaitable<bool> MoveNextAsync() =>
-                _enumerator.MoveNextAsync().ConfigureAwait(_continueOnCapturedContext);
-
-            public T Current => _enumerator.Current;
-
-            public ConfiguredValueTaskAwaitable DisposeAsync() =>
-                _enumerator.DisposeAsync().ConfigureAwait(_continueOnCapturedContext);
-        }
-    }
-}
